Introduced usage of ASN1AdaptiveTypeValues. An adaptive type is supposed to always carry the type it operates on. Some operations take the type into account to provide extra service. Reading and writing fields throw exceptions instead of returning nil. An adaptive type can be self expanding, meaning it can add sub structures automatically when message chaining is used. The current approach is now in testing and might change.
Unified read accessors in ASN1 for reading in structures. Added an argument for the class to write values to. This way adaptive derived objects can add more special behaviour to the parsing.

Changed read-write visitor behaviour. Removed the use of a stack collection to track nested structure. Now uses a variable in the compiled method to cache the former value. Avoids going through a collection to access the latest object. This should give a speedup