Remove doublePeek and move to peekTwice
PetitParser version 1.8-snapshot has been removed and this forces us to move. I
decided to move to version 1.9 which introduces PPContext, which forces us to
move to modify doublePeek.. and at that time we can use peekTwice which looks
to be identical. Remove >>#doublePeek, rename the variables of the only caller
and remove the unused variable.