Tags >= 31 need to be encoded in multiple octets
BERIdentifier was capable of decoding such tags but not to encode those. Add a simple unit test that makes a roundtrip test and extend ASN1Tag to write itself onto a stream and to emit multiple bytes.

I think it would be better to move the writing to BERIdentifier directly and then have BERStreamFormat know about reading/writing such tags.

The issue was found while trying to write the sgsn-Number in
a InsertSubscriberDataArgV3.