ASN1BitStringValue/ASN1BitStringType encoding fixes
The >>#encode: routine assumed it could just trim bits to the right. This is
wrong because on types like ODB-GeneralData a minimum size constraint is
set and by trimming this would be violated.

At the same time ASN1BitStringType does not know (anymore?) about this
constraint.

For now change ASN1BitStringValue to pad the string to the maximum known
amount of bits. This makes us use a lot more bits/bytes in cases where we
could use fewer.

Update the test cases to use >>#assert:equals: and the test result to match
the current one (non truncated view).